fix dto atributos

+++ src/main/java/com/ELShovi/dto/dashboard/VentasPorUsuarioDTO.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class VentasPorUsuarioDTO {
     private String usuario;
-    private Double total;
+    private BigDecimal total;
 }
